refactor MiniAdventure
