+++ package.json
@@ -17,7 +17,7 @@
     "proj4": "2.17.0",
     "react": "19.2.3",
     "react-dom": "19.2.3",
-    "react-router": "7.12.0"
+    "react-router": "7.10.1"
   },
   "devDependencies": {
     "@babel/preset-react": "7.28.5",
@@ -54,7 +54,7 @@
       "proj4": "2.17.0",
       "react": "19.2.3",
       "react-dom": "19.2.3",
-      "react-router": "7.12.0"
+      "react-router": "7.10.1"
     },
     "devDependencies": {
       "@babel/preset-react": "7.28.5",
