+++ src/main/java/com/pdg/adventure/MiniAdventureContent.java
@@ -1,0 +1,353 @@
+package com.pdg.adventure;
+
+import lombok.Getter;
+
+import java.util.Map;
+
+import com.pdg.adventure.api.Command;
+import com.pdg.adventure.api.Container;
+import com.pdg.adventure.api.PreCondition;
+import com.pdg.adventure.server.action.*;
+import com.pdg.adventure.server.condition.CarriedCondition;
+import com.pdg.adventure.server.condition.NotCondition;
+import com.pdg.adventure.server.condition.WornCondition;
+import com.pdg.adventure.server.engine.GameContext;
+import com.pdg.adventure.server.location.GenericDirection;
+import com.pdg.adventure.server.location.Location;
+import com.pdg.adventure.server.parser.GenericCommand;
+import com.pdg.adventure.server.parser.GenericCommandDescription;
+import com.pdg.adventure.server.storage.message.MessagesHolder;
+import com.pdg.adventure.server.support.DescriptionProvider;
+import com.pdg.adventure.server.support.Variable;
+import com.pdg.adventure.server.support.VariableProvider;
+import com.pdg.adventure.server.tangible.GenericContainer;
+import com.pdg.adventure.server.tangible.Item;
+import com.pdg.adventure.server.vocabulary.Vocabulary;
+import com.pdg.adventure.model.Word;
+import com.pdg.adventure.server.AdventureConfig;
+
+public class MiniAdventureContent {
+    private static final String SMALL_TEXT = "small";
+
+    private final CommandFactory commandFactory;
+    private final MessagesHolder allMessages;
+    private final Vocabulary allWords;
+    private final VariableProvider variableProvider;
+    private final Map<String, Location> allLocations;
+    private final GameContext gameContext;
+
+    @Getter
+    private final Container allItems;
+    @Getter
+    private Location portal;
+    @Getter
+    private Location location;
+    @Getter
+    private Location house;
+    @Getter
+    private Container pocket;
+
+    public MiniAdventureContent(AdventureConfig adventureConfig, GameContext gameContext, CommandFactory commandFactory) {
+        this.commandFactory = commandFactory;
+        this.gameContext = gameContext;
+        this.allMessages = adventureConfig.allMessages();
+        this.allWords = adventureConfig.allWords();
+        this.allLocations = adventureConfig.allLocations();
+        this.variableProvider = adventureConfig.allVariables();
+        this.allItems = new GenericContainer(new DescriptionProvider("all items"), 9999);
+    }
+
+    public void setUp() {
+        setUpMessages();
+        setUpVocabulary();
+        gameContext.tell("You have words!");
+
+        setUpVariables();
+        gameContext.tell("You have variables!");
+
+        setUpPocket();
+
+        setUpLocations();
+        gameContext.tell("You have places!");
+
+        setUpItems();
+        gameContext.tell("You have items!");
+
+        setUpDirections();
+
+        setUpItemsInFirstLocation(location);
+        setUpItemsInPortal(portal);
+        setUpItemsInHut(house);
+        gameContext.tell("You have items in places!");
+    }
+
+    public void setUpMessages() {
+        allMessages.addMessage("-13", "You already carry %s.");
+        allMessages.addMessage("-12", "A %s appears in the %s.");
+        allMessages.addMessage("-11", "The %s evaporates into thin air.");
+        allMessages.addMessage("-10", "You carry:");
+        allMessages.addMessage("-9", "You put %s into %s.");
+        allMessages.addMessage("-8", "The %s is full.");
+        allMessages.addMessage("-7", "You can't remove %s.");
+        allMessages.addMessage("-6", "You can't wear %s.");
+
+        allMessages.addMessage("-2", "I don't understand, please rephrase.");
+        allMessages.addMessage("-1", "You can't do that.");
+        allMessages.addMessage("0", "OK.");
+
+        allMessages.addMessage("1", "As you inspect the ring you notice the shape of a portal engraved in it.");
+        allMessages.addMessage("2", "You skin the rabbit and are left with a rabbit pelt.");
+        allMessages.addMessage("3", "You need a knife.");
+        allMessages.addMessage("4", "You enter the game.");
+        allMessages.addMessage("5", "This is the first location.");
+        allMessages.addMessage("6", """
+                You find yourself in a field of lush grass and colourful flowers surrounding a small hut. It looks too
+                beautiful to be true, much more like a painting.
+                Suddenly, you notice a glowing portal!""");
+        allMessages.addMessage("7", "The flowers look beautiful.");
+        allMessages.addMessage("8", "You are in a small portal.");
+        allMessages.addMessage("9", "The portal slowly fades away already, it looks like it closes soon!");
+        allMessages.addMessage("10", "You are in a small brick hut.");
+        allMessages.addMessage("11", "This is a small hut made of bricks. There is nothing in here.");
+        allMessages.addMessage("12", "There seems to be an invisible barrier. You need some sort of key to enter " +
+                                     "the portal.");
+        allMessages.addMessage("14", "pair of leathery gloves");
+        allMessages.addMessage("15", "These gloves would withstand any harsh treatment!");
+        allMessages.addMessage("16", "a small knife");
+        allMessages.addMessage("17", "The knife is exceptionally sharp. Don't cut yourself!");
+        allMessages.addMessage("18", "The knife is too sharp! You need to wear some gloves.");
+        allMessages.addMessage("19", "You may sell it to a trader, but that is part of a different adventure.");
+        allMessages.addMessage("20", "You may cook it, but that is part of a different adventure");
+        allMessages.addMessage("21", "The rabbit looks very tasty!");
+        allMessages.addMessage("22", "The rabbit hops away.");
+    }
+
+    private void setUpVocabulary() {
+        allWords.createNewWord("quit", Word.Type.VERB);
+        allWords.createSynonym("exit", "quit");
+        allWords.createSynonym("bye", "quit");
+        allWords.createNewWord("save", Word.Type.VERB);
+        allWords.createNewWord("load", Word.Type.VERB);
+        allWords.createNewWord("inventory", Word.Type.VERB);
+        allWords.createSynonym("i", "inventory");
+        allWords.createNewWord("north", Word.Type.VERB);
+        allWords.createSynonym("n", "north");
+        allWords.createNewWord("east", Word.Type.VERB);
+        allWords.createSynonym("e", "east");
+        allWords.createNewWord("south", Word.Type.VERB);
+        allWords.createSynonym("s", "south");
+        allWords.createNewWord("west", Word.Type.VERB);
+        allWords.createSynonym("w", "west");
+        allWords.createNewWord("leave", Word.Type.VERB);
+        allWords.createNewWord("describe", Word.Type.VERB);
+        allWords.createSynonym("look", "describe");
+        allWords.createSynonym("l", "describe");
+        allWords.createSynonym("desc", "describe");
+        allWords.createSynonym("examine", "describe");
+        allWords.createSynonym("inspect", "describe");
+
+        allWords.createNewWord("knife", Word.Type.NOUN);
+        allWords.createNewWord("pelt", Word.Type.NOUN);
+
+        allWords.createNewWord("get", Word.Type.VERB);
+        allWords.createSynonym("take", "get");
+        allWords.createNewWord("drop", Word.Type.VERB);
+
+        allWords.createNewWord("open", Word.Type.VERB);
+        allWords.createNewWord("enter", Word.Type.VERB);
+        allWords.createSynonym("go", "enter");
+
+        allWords.createNewWord("cut", Word.Type.VERB);
+        allWords.createSynonym("kill", "cut");
+        allWords.createSynonym("stab", "cut");
+
+        allWords.createNewWord("wear", Word.Type.VERB);
+        allWords.createNewWord("remove", Word.Type.VERB);
+
+        allWords.createNewWord("rabbit", Word.Type.NOUN);
+        allWords.createSynonym("hare", "rabbit");
+
+        allWords.createNewWord("ring", Word.Type.NOUN);
+        allWords.createNewWord("flowers", Word.Type.NOUN);
+
+        allWords.createNewWord("big", Word.Type.ADJECTIVE);
+        allWords.createNewWord("skinned", Word.Type.ADJECTIVE);
+        allWords.createNewWord(SMALL_TEXT, Word.Type.ADJECTIVE);
+        allWords.createNewWord("golden", Word.Type.ADJECTIVE);
+
+        allWords.createNewWord("portal", Word.Type.NOUN);
+        allWords.createNewWord("hut", Word.Type.NOUN);
+        allWords.createSynonym("house", "hut");
+
+        allWords.createNewWord("leathery", Word.Type.ADJECTIVE);
+        allWords.createNewWord("gloves", Word.Type.NOUN);
+
+        allWords.createNewWord("here", Word.Type.NOUN);
+
+        allWords.createNewWord("help", Word.Type.VERB);
+    }
+
+    private void setUpVariables() {
+        variableProvider.set(new Variable("wornRing", "false"));
+    }
+
+    private void setUpPocket() {
+        pocket = new GenericContainer(new DescriptionProvider("your pocket"), 5);
+    }
+
+    private void setUpLocations() {
+        DescriptionProvider locationDescription = new DescriptionProvider("first", "location");
+        locationDescription.setShortDescription(allMessages.getMessage("5"));
+        locationDescription.setLongDescription(allMessages.getMessage("6"));
+        location = new Location(locationDescription, pocket);
+        commandFactory.setUpLookCommands(location);
+
+        GenericCommandDescription flowerCommandDescription = new GenericCommandDescription("describe", "flowers");
+        GenericCommand checkFlowerCommand = new GenericCommand(flowerCommandDescription,
+                                                               new MessageAction(allMessages.getMessage("7"),
+                                                                                 allMessages));
+        location.addCommand(checkFlowerCommand);
+
+        DescriptionProvider portalDescription = new DescriptionProvider("fading", "portal");
+        portalDescription.setShortDescription(allMessages.getMessage("8"));
+        portalDescription.setLongDescription(allMessages.getMessage("9"));
+        portal = new Location(portalDescription, pocket);
+        commandFactory.setUpLookCommands(portal);
+
+        DescriptionProvider houseDescription = new DescriptionProvider(SMALL_TEXT, "hut");
+        houseDescription.setShortDescription(allMessages.getMessage("10"));
+        houseDescription.setLongDescription(allMessages.getMessage("11"));
+        house = new Location(houseDescription, pocket);
+        commandFactory.setUpLookCommands(house);
+
+        allLocations.put(location.getId(), location);
+        allLocations.put(portal.getId(), portal);
+        allLocations.put(house.getId(), house);
+    }
+
+    private void setUpItems() {
+        // gloves
+        Item gloves = new Item(new DescriptionProvider("gloves"), true);
+        allItems.add(gloves);
+        gloves.setShortDescription(allMessages.getMessage("14"));
+        gloves.setLongDescription(allMessages.getMessage("15"));
+        commandFactory.setUpWearCommands(gloves);
+        commandFactory.setUpLookCommands(gloves);
+        commandFactory.setUpTakeCommands(gloves);
+
+        // knife
+        Item knife = new Item(new DescriptionProvider(SMALL_TEXT, "knife"), true);
+        allItems.add(knife);
+        knife.setShortDescription(allMessages.getMessage("16"));
+        knife.setLongDescription(allMessages.getMessage("17"));
+        GenericCommand getNotSuccessful = new GenericCommand(new GenericCommandDescription("get", knife),
+                                                             new MessageAction(allMessages.getMessage("18"),
+                                                                               allMessages));
+        PreCondition glovesWorn = new WornCondition((gloves));
+        NotCondition glovesNotWorn = new NotCondition(glovesWorn);
+        getNotSuccessful.addPreCondition(glovesNotWorn);
+        knife.addCommand(getNotSuccessful);
+        commandFactory.setUpLookCommands(knife);
+        commandFactory.setUpTakeCommands(knife);
+
+        // rabbit
+        Item pelt = new Item(new DescriptionProvider(SMALL_TEXT, "pelt"), true);
+        allItems.add(pelt);
+        pelt.setLongDescription(allMessages.getMessage("19"));
+        commandFactory.setUpLookCommands(pelt);
+        commandFactory.setUpTakeCommands(pelt);
+
+        Item skinnedRabbit = new Item(new DescriptionProvider("skinned", "rabbit"), true);
+        allItems.add(skinnedRabbit);
+        skinnedRabbit.setLongDescription(allMessages.getMessage("20"));
+        commandFactory.setUpLookCommands(skinnedRabbit);
+        commandFactory.setUpTakeCommands(skinnedRabbit);
+
+        Item rabbit = new Item(new DescriptionProvider(SMALL_TEXT, "rabbit"), true);
+        allItems.add(rabbit);
+        rabbit.setLongDescription(allMessages.getMessage("21"));
+        GenericCommand cutNotSuccessfully = new GenericCommand(new GenericCommandDescription("cut", rabbit),
+                                                               new MessageAction(allMessages.getMessage("3"),
+                                                                                 allMessages));
+        CarriedCondition knifeCarried = new CarriedCondition(knife, gameContext);
+        NotCondition knifeNotCarried = new NotCondition(knifeCarried);
+        cutNotSuccessfully.addPreCondition(knifeNotCarried);
+        rabbit.addCommand(cutNotSuccessfully);
+
+        GenericCommand cutSuccessfully = new GenericCommand(new GenericCommandDescription("cut", rabbit),
+                                                            new MessageAction(allMessages.getMessage("2"),
+                                                                              allMessages));
+        cutSuccessfully.addPreCondition(knifeCarried);
+        cutSuccessfully.addFollowUpAction(new CreateAction(pelt, rabbit::getParentContainer, allMessages));
+        cutSuccessfully.addFollowUpAction(new CreateAction(skinnedRabbit, location::getItemContainer, allMessages));
+        cutSuccessfully.addFollowUpAction(new DestroyAction(rabbit, allMessages));
+        rabbit.addCommand(cutSuccessfully);
+        commandFactory.setUpLookCommands(rabbit);
+        commandFactory.setUpTakeCommands(rabbit, new MessageAction(allMessages.getMessage("22"), allMessages));
+
+        // ring
+        DescriptionProvider ringDescription = new DescriptionProvider("golden", "ring");
+        ringDescription.setLongDescription(allMessages.getMessage("1"));
+        Item ring = new Item(ringDescription, true);
+        allItems.add(ring);
+        commandFactory.setUpWearCommands(ring);
+        commandFactory.setUpLookCommands(ring);
+        commandFactory.setUpTakeCommands(ring);
+    }
+
+    private void setUpDirections() {
+        Item ring = (Item) allItems.findItemByShortDescription("golden", "ring");
+
+        GenericCommandDescription enterPortalCommandDescription = new GenericCommandDescription("enter", portal);
+        GenericCommand enterPortalCommand = new GenericCommand(enterPortalCommandDescription,
+                                                               new MovePlayerAction(portal, allMessages, gameContext));
+        enterPortalCommand.addPreCondition(new WornCondition(ring));
+
+        Command enterCommand2 = new GenericCommand(enterPortalCommandDescription,
+                                                   new MessageAction(allMessages.getMessage("12"), allMessages));
+        enterCommand2.addPreCondition(new NotCondition(new WornCondition(ring)));
+
+        GenericDirection toPortal = new GenericDirection(allLocations, enterCommand2, portal.getId(), true);
+        toPortal.addCommand(enterPortalCommand);
+
+        commandFactory.setUpLookCommands(toPortal);
+        location.addDirection(toPortal);
+
+        GenericCommandDescription enterHouseCommandDescription = new GenericCommandDescription("enter", house);
+        GenericCommand enterHouseCommand = new GenericCommand(enterHouseCommandDescription,
+                                                              new MovePlayerAction(house, allMessages, gameContext));
+        GenericDirection toHouse = new GenericDirection(allLocations, enterHouseCommand, house.getId(), true);
+
+        commandFactory.setUpLookCommands(toHouse);
+        location.addDirection(toHouse);
+
+        GenericCommandDescription leavePortalCommandDescription = new GenericCommandDescription("leave", location);
+        GenericCommand leaveCommand = new GenericCommand(leavePortalCommandDescription,
+                                                         new MovePlayerAction(location, allMessages, gameContext));
+        GenericDirection toLocation = new GenericDirection(allLocations, leaveCommand, location.getId());
+        portal.addDirection(toLocation);
+
+        GenericCommandDescription leaveHouseCommandDescription = new GenericCommandDescription("north", location);
+        leaveCommand = new GenericCommand(leaveHouseCommandDescription, new MovePlayerAction(location, allMessages, gameContext));
+        toLocation = new GenericDirection(allLocations, leaveCommand, location.getId());
+        house.addDirection(toLocation);
+    }
+
+    private void setUpItemsInFirstLocation(Location aLocation) {
+        Item rabbit = (Item) allItems.findItemByShortDescription(SMALL_TEXT, "rabbit");
+        aLocation.addItem(rabbit);
+
+        Item ring = (Item) allItems.findItemByShortDescription("golden", "ring");
+        aLocation.addItem(ring);
+    }
+
+    private void setUpItemsInPortal(Location aLocation) {
+        Item knife = (Item) allItems.findItemByShortDescription(SMALL_TEXT, "knife");
+        aLocation.addItem(knife);
+    }
+
+    private void setUpItemsInHut(Location aLocation) {
+        Item gloves = (Item) allItems.findItemByShortDescription("", "gloves");
+        aLocation.addItem(gloves);
+    }
+}
